back/requirements.txt
Pillow==8.3.2
SQLAlchemy
python-dotenv
python-multipart
psycopg2-binary
numpy==1.25.0
pandas==1.5.2
uvicorn==0.20.0
fastapi==0.89.1
passlib[bcrypt]
python-jose[cryptography]